app/requirements.txt
folium==0.16.0
numpy==1.26.4
pandas==2.2.2
phonenumbers==8.13.34
Pillow==10.3.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
streamlit==1.33.0
streamlit_authenticator==0.3.2
streamlit-card==1.0.2
streamlit_folium==0.19.1
tqdm==4.66.1
google-auth==2.26.2
google-cloud-bigquery==3.16.0
google-cloud-core==2.4.1
bcrypt==4.1.3
db-dtypes==1.2.0
openpyxl==3.1.2